final implementation of project3

+++ Project3/src/Cecs475.BoardGames.ComputerOpponent/MinimaxAi.cs
@@ -16,7 +16,7 @@ namespace Cecs475.BoardGames.ComputerOpponent
     public class MinimaxAi : IGameAi
     {
         private long mMaxDepth;
-        public MinimaxAi(long maxDepth)
+        public MinimaxAi(int maxDepth)
         {
             mMaxDepth = maxDepth;
         }
